+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2] - 2021-03-07
+### Added
+- PutItemWithContext() implemented
+- QueryWithContext() implemented
+
 ## [0.0.1] - 2021-02-14
 ### Added
 - Query() implemented